% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelReSample-package.R
\docType{package}
\name{KernelReSample-package}
\alias{KernelReSample}
\alias{KernelReSample-package}
\title{KernelReSample: Kernel Resampling Estimators}
\description{
Implements kernel ridge regression (KRR) with resampling-based prediction.
}
\details{
Main functions:
\itemize{
\item \code{\link{kr_boot}}: Baseline K-means resampling estimator.
\item \code{\link{kr_boot_plus}}: Advanced estimator with multiple kernels,
clustering backends, and bandwidth strategies.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ykong21/KernelReSample}
  \item Report bugs at \url{https://github.com/Ykong21/KernelReSample/issues}
}

}
\author{
\strong{Maintainer}: Yundi Kong \email{yundikong21@gmail.com}

}
\keyword{internal}
